Speed up problem generation by introducing aux variables
Introduce auxiliary variables that are equal to boolean
expressions of other variables that are used more than once. This
increases the number of variables by a negligible amount but
greatly decreases the number of terms in constraints.

The model generation is so much faster now, that there is no
need for the "prepare" script any more.